Entity creada Dao Creado Database Creada Repositorio empezando el ViewModel ViewModel terminando itemlist arreglado adapter recyclerView instancia el adaptador y mostrar datos guardado se moeve botoncito del segundo frgmaento al primer fragmento segundo fragmento ya graba en el RecyclerView Edición de datos falta borrar los datos Bundle del primer al segundo fragento editando datos en el segundo fragmento Lista la edición, validación vacios. Falta calcular y eliminar.

+++ app/src/main/java/com/example/registrodeconsumo/SecondFragment.kt
@@ -58,22 +58,22 @@ class SecondFragment : Fragment() {
 
 
             val textseg = iditemTv.text.toString()
-            val precioseg = precioTv.text.toString().toInt()
-            val cantseg = cantTv.text.toString().toInt()
+            val precioseg = precioTv.text.toString()
+            val cantseg = cantTv.text.toString()
 
 
             if (!textseg.isEmpty()) {
                 if (idpedidos != null) {
                     val mPedidos = Pedidos(
                         item = textseg,
-                        precio = precioseg,
-                        cantidad = cantseg,
+                        precio = precioseg.toInt(),
+                        cantidad = cantseg.toInt(),
                         id = idpedidos!!
                     )
                     mViewModel.updatePedidos(mPedidos)
                 } else {
 
-                    val mPedidos = Pedidos(item = textseg, precio = precioseg, cantidad = cantseg)
+                    val mPedidos = Pedidos(item = textseg, precio = precioseg.toInt(), cantidad = cantseg.toInt())
                     mViewModel.insertPedidos(mPedidos)
                 }
             }else{
@@ -81,5 +81,11 @@ class SecondFragment : Fragment() {
             }
             findNavController().navigate(R.id.action_SecondFragment_to_FirstFragment)
         }
+
+        ///Termina el botón guardar
+
+
+
+
     }
 }
